rename src/omnimock.rs to src/expectation.rs
It just took me awhile to figure out what to call this file.
